final/frontend/requirements.txt
numpy>=1.22.4,<2.0
streamlit>=1.32.0
plotly
pandas
requests
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
